Fix - resources tab in a cell is now int[] Fix - change print_line/row parameters : uint32 -> int32 Fix - add UNUSED to unused parameters Fix - passing pointer to struct pos in care_positions function Fix - get_resource_on_cell returning int instead of size_t Fix - change function related to incantation to static functions Fix - List initialisation in connect_nbr

+++ sources/server/commands/print_content.c
@@ -11,35 +11,35 @@
 static void print_cell_content(cell_t *cell, client_t *client);
 
 void print_line_content(map_t *map_infos, client_t *client,
-	uint32_t y, uint32_t tiles)
+	int32_t y, uint32_t tiles)
 {
 	cell_t *tmp;
-	uint32_t x = client->infos->pos.x  - (tiles / 2);
+	int32_t x = client->infos->pos.x  - (tiles / 2);
 
-	if (y < 0 || y >= map_infos->y)
+	if (y < 0 || (uint32_t)y >= map_infos->y)
 		y = (y + map_infos->y) % map_infos->y;
 	for (uint32_t j = 0; j < tiles; j++) {
-		if (x < 0 || x >= map_infos->x)
+		if (x < 0 || (uint32_t)x >= map_infos->x)
 			x = (x + map_infos->x) % map_infos->x;
-		tmp = get_cell_at(map_infos, x, y);
+		tmp = get_cell_at(map_infos, (uint32_t)x, (uint32_t)y);
 		print_cell_content(tmp, client);
 		x++;
 	}
 }
 
 void print_row_content(map_t *map_infos, client_t *client,
-	uint32_t x, uint32_t tiles)
+	int32_t x, uint32_t tiles)
 {
 	cell_t *tmp;
-	uint32_t y = client->infos->pos.y  - (tiles / 2);
+	int32_t y = client->infos->pos.y  - (tiles / 2);
 
-	if (x < 0 || x >= map_infos->x)
+	if (x < 0 || (uint32_t)x >= map_infos->x)
 		x = (x + map_infos->x) % map_infos->x;
 	for (uint32_t j = 0; j < tiles; j++) {
-		if (y < 0 || y >= map_infos->y) {
+		if (y < 0 || (uint32_t)y >= map_infos->y) {
 			y = (y + map_infos->y) % map_infos->y;
 		}
-		tmp = get_cell_at(map_infos, x, y);
+		tmp = get_cell_at(map_infos, (uint32_t)x, (uint32_t)y);
 		print_cell_content(tmp, client);
 		y++;
 	}
